fix(#6133): move htons tests into their own file, SyscallTest.java is at the line ceiling
SyscallTest.java is already at Checkstyle's 1000-line file-length limit;
adding the htons tests there tipped it over. A dedicated HtonsTest.java
keeps the coverage without growing an already-maxed-out file.
